Move replicache.dart to src
This is in preparation of hiding forTesting
